Reorganize test utilities and remove dead code
- Created proper src/test-utils/ directory structure with:
  - temp-directory.ts: Temporary directory utilities
  - mock-session.ts: Session and context mocking utilities
- Removed unused dead code from tools/test-utils.ts:
  - createTestToolCall: Never imported, tests use local implementations
  - createToolCallFactory: Never imported anywhere
- Updated all imports to use the new organized structure
- Improved separation of concerns: general test utilities are no longer buried in tools/

This creates better architecture with test utilities in their proper location and eliminates unused helper functions that were never adopted.
