+++ src/tools/executor.test.ts
@@ -7,7 +7,8 @@ import { FileReadTool } from '~/tools/implementations/file-read';
 import { ApprovalCallback, ApprovalDecision, ApprovalPendingError } from '~/tools/approval-types';
 import { writeFile } from 'fs/promises';
 import { join } from 'path';
-import { createTestTempDir, createMockToolContext } from '~/tools/test-utils';
+import { createTestTempDir } from '~/test-utils/temp-directory';
+import { createMockToolContext } from '~/test-utils/mock-session';
 import { existsSync } from 'fs';
 import { clearProcessTempDirCache } from '~/config/lace-dir';
 import { z } from 'zod';
